Allow custom theme install
Adds a setting custom-theme which operats similar to ubuntu-theme and
default-theme. The provided resource folder is placed in the themes folder
and apache is setup to serve static content for the theme. These leaves
the default theme untouched allowing the theme to override files based
on the build in horizon theme capabilities. For details on theming
capabilities see: https://docs.openstack.org/horizon/latest/configuration/themes.html

Change-Id: I91ad19e8aad5c0e0773d42fa4f085cbcecb82458